pyproject.toml
[tool.poetry]
name = "mulesoft-accenture"
version = "0.1.0"
description = "Multi-agent system for business opportunities using ADK"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "teams_agent" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
google-adk = "^1.12.0"
google-generativeai = "^0.8.5"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
